update actions
delete Snapshot action, Snapshots AND normal versions will automatically be build on push to master (according to the version in the pom.xml)
verify-action is only triggered on branches besides the master branch. Pushing on master branch automatically triggers the deploy action that also verifies the build.

Also update to JDK 14